requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
gTTS==2.5.1
aiofiles==23.2.1
apscheduler==3.10.4
sqlalchemy==2.0.32
aiosqlite==0.19.0
